% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{convolution}
\alias{convolution}
\title{Convoluzione}
\usage{
convolution(n, lambda, shape, scale)
}
\arguments{
\item{n}{int}

\item{lambda}{int}

\item{shape}{double}

\item{scale}{double}
}
\value{
double vector of length n
}
\description{
calcolo della convoluzione totale
}
\examples{
convolution(n=5, lambda = 100, shape = 2, scale = 24)
}
