% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot.ritardo_giuseppe}
\alias{plot.ritardo_giuseppe}
\title{Plot method per la classe ritardo_giuseppe}
\usage{
\method{plot}{ritardo_giuseppe}(x, ...)
}
\arguments{
\item{x}{object of class ritardo_giuseppe}

\item{...}{extra parameter for geom_histogram}
}
\value{
invisible(NULL). Plot histogram as side effect
}
\description{
plot method per oggetti classe ritardo_giuseppe
}
\examples{
ritardo_g <- ritardo_giuseppe()
plot(ritardo_g, fill = 'green', col = 'darkgreen')
}
