% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{print.ritardo_giuseppe}
\alias{print.ritardo_giuseppe}
\title{Print method per la classe ritardo_giuseppe}
\usage{
\method{print}{ritardo_giuseppe}(x)
}
\arguments{
\item{x}{object of class ritardo_giuseppe}
}
\value{
invisible(NULL).
}
\description{
print specifico per oggetti classe ritardo_giuseppe
}
\examples{
ritardo_g <- ritardo_giuseppe()
print(ritardo_g)
}
