% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ritardo_giuseppe}
\alias{ritardo_giuseppe}
\title{ritardo di giuseppe}
\usage{
ritardo_giuseppe(lambda = 100, shape = 2, scale = 24, n = 10^5,
  p = 0.999)
}
\arguments{
\item{lambda}{int default 100}

\item{shape}{double default 2}

\item{scale}{double default 24}

\item{n}{int default 10^5}

\item{p}{double default 0.999}
}
\value{
list of three elements:
\itemize{
\item conv: convolution vector
\item ritardo_atteso: ritardo medio
\item ritardo_a_rischio: quantile del ritardo di Giuseppe
}
}
\description{
calcolo del ritardo di
}
\examples{
ritardo_g <- ritardo_giuseppe()
ritardo_g$ritardo_a_rischio
ritardo_g$ritardo_atteso
head(ritardo_g$conv)
class(ritardo_g)
}
