% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{single_convolution}
\alias{single_convolution}
\title{single convolution}
\usage{
single_convolution(np_i, shape, scale)
}
\arguments{
\item{np_i}{int frequenza ritardo}

\item{shape}{double}

\item{scale}{double}
}
\value{
double ritardo comulato in un anno
}
\description{
calcolo della convoluzione
}
\examples{
single_convolution(np_i = 100, shape = 2, scale = 24)
}
