% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{single_convolution_i}
\alias{single_convolution_i}
\title{single convolution on the ith np}
\usage{
single_convolution_i(i, np, shape, scale)
}
\arguments{
\item{i}{int i-esima componente}

\item{shape}{double}

\item{scale}{double}

\item{np_i}{int frequenza ritardo}
}
\value{
double ritardo comulato in un anno
}
\description{
calcolo della convoluzione sull'i-esimo np
}
\examples{
single_convolution_i(i=2, np = c(100, 120, 140), shape = 2, scale = 24)
}
